pyproject.toml
[project]
name = "pelvibiz-agent-api"
version = "0.1.0"
description = "PelviBiz AI Hub Agent API — Intelligent carousel generation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "supabase>=2.12.0",
    "httpx>=0.28.0",
    "openai>=1.60.0",
    "python-multipart>=0.0.18",
    "google-genai>=1.0.0",
    "Pillow>=11.0.0",
    "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]
